chore: new networkURL prop, move WelcomeVersion
